move generate_markdown to seperate file, unify retfetching of test_files
