+++ test/utils.py
@@ -1,0 +1,33 @@
+import glob
+import os
+import re
+import sys
+
+
+def collect_test_files():
+    """
+    Option -> Example
+    Test files for a rule -> 'python3 test_main.py alb001'
+    Also applies for multiple rules -> 'python3 test_main.py alb001 alb002'
+    Test files for a single file -> 'python3 test_main.py <path>.ifc
+    Also applies for multiple files -> 'python3 test_main.py <path1>.ifc <path2>.ifc'
+    Codes and rules can also be combined -> 'python3 test_main.py alb001 <path>.ifc'
+    """
+    #todo @gh check for removed testfiles in documentation
+    args = [a for a in sys.argv[1:] if not a.startswith('-')]
+    rule_code_pattern = re.compile(r"^[a-zA-Z]{3}\d{3}$")
+    rule_codes = list(filter(lambda arg: rule_code_pattern.match(arg), args))
+
+    test_files = []
+    for code in rule_codes:
+        paths = glob.glob(os.path.join(os.path.dirname(__file__), "files/", code.lower(), "*.ifc"))
+        if not paths:
+            print(f"No IFC files were found for the following rule code: {code}. Please provide test files or verify the input.")
+        test_files.extend(paths)
+
+    file_pattern =  r".*\.ifc(\')?$" #matches ".ifc" and "ifc'"
+    test_files.extend([s.strip("'") for s in args if re.match(file_pattern, s)])
+
+    if not args: # for example, with 'pytest -sv'
+        test_files = glob.glob(os.path.join(os.path.dirname(__file__), "files/**/*.ifc"), recursive=True)
+    return test_files
